requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv
qdrant-client
openai
uuid